requirements.txt
fastapi>=0.111.0,<1.0.0
uvicorn[standard]>=0.29.0,<1.0.0
web3>=6.18.0,<7.0.0
requests>=2.32.3,<3.0.0
httpx>=0.27.0,<1.0.0
ccxt>=4.4.30,<5.0.0
python-dotenv>=1.0.1,<2.0.0
APScheduler>=3.10.4,<4.0.0


